requirements.txt
playwright>=1.58.0
playwright-stealth>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
